Listen for crash event and respond appropriately
If there was a browser crash, the runner would just hang. Now,
we catch the event and shutdown gracefully.

Also changed the error print from chromedp.Run to a non-fatal one
because we need to gracefully shutdown the server. And that is orthogonal
to whether chromedp had an error or not.