fix parsing's use of flag
The flag package's global obscurred the fact the last commit had mixed
instances of the FlagSet being used in the loop along with the
flag.CommandLine global. And when using it with the CommandLine FlagSet,
it may not matter. Only if this function is used somewhere else with a
different FlagSet would this problem have surfaced.