replace flag.Parse with gentleParse (#34)
* replace flag.Parse with gentleParse

gentleParse does away with having to call out all the arguments that
will be passed to the wasm image.

Now the wasm image will be passed all the arguments that the tool
doesn't recognize itself.

Also there are unit test examples of the parsing.

* fix parsing's use of flag

The flag package's global obscurred the fact the last commit had mixed
instances of the FlagSet being used in the loop along with the
flag.CommandLine global. And when using it with the CommandLine FlagSet,
it may not matter. Only if this function is used somewhere else with a
different FlagSet would this problem have surfaced.

* Fix and cleanup gentleParse

Fixes the case of an unrecognized argument using "=".

Also fixes case where flag.Parse ate the "--" because it had recognized everything up
to that point; we want the "--" passed back to the caller.

Adds some tests for the two aspects just fixed.

Moved gentleParse to parse.go.

Renamed examples in parse_test.go to contain "Parse" so test runs
like 'go test -run Parse -v' are possible to hit just those tests.

* Cleanup parse_test.go

The unit tests for gentleParse are converted from Example functions
to a *testing.T framework where a list of args and a list of expected
result lines are given. They are still kept as separate tests to allow
running some subset when wanting to debug something.