Do not cancel context on exception
Sometimes, exception can happen if a callback runs
after the test has exited. In that case, cancelling
a context fails the test which is not intentional.

Since cancelling happens anyways after a test exits,
there is not much point manually cancelling it
and we try to keep parity with non-browser environments.

Fixes #60